% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynofits.R
\name{aic_bic}
\alias{aic_bic}
\title{Compare different models using AIC/BIC}
\usage{
aic_bic(...)
}
\arguments{
\item{...}{One or more dynamic copula models to compare. They can be comma
seperated or in a list.}
}
\description{
Compare different models using AIC/BIC
}

