% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynohelpers.R
\name{aprx}
\alias{aprx}
\title{Approximate small sample p-value for the generalized likelihood test}
\usage{
aprx(x, np, N)
}
\arguments{
\item{x}{Test statistic from \code{\link{BreakPointTestStatistic}}}

\item{np}{The number of parameters for the chosen copula model}

\item{N}{The length of the time series}
}
\value{
A p-value
}
\description{
Approximate small sample p-value for the generalized likelihood test
}

