% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynohelpers.R
\name{autoBPtest}
\alias{autoBPtest}
\title{For the sequential breakpoint model, a closure that recursively searchs for
all statistically significant breakpoints.}
\usage{
autoBPtest(x, y, f1, f2 = NULL, parallel = FALSE, date_names = NULL,
  ncores)
}
\arguments{
\item{f1}{An integer representing the family of the component copula applied
to u. This is passed to \code{\link{cop_static}}}

\item{f2}{An integer representing the family of the component copula applied
to v. This is passed to \code{\link{cop_static}}}

\item{parallel}{Logical switch to run the breakpoint search in parallel.}

\item{date_names}{Optional vector of date names}

\item{u}{A vector of uniform marginal values.}

\item{v}{A vector of uniform marginal values.}
}
\description{
For the sequential breakpoint model, a closure that recursively searchs for
all statistically significant breakpoints.
}

