% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynohelpers.R
\name{BreakAnalysis}
\alias{BreakAnalysis}
\title{For the sequential breakpoint model, search for a breakpoint in a subset of 
values for two given marginal distributions#'}
\usage{
BreakAnalysis(u, v, series, fam1, fam2 = NULL, date_names = NULL,
  parallel = FALSE, ncores, cluster)
}
\arguments{
\item{u}{A vector of uniform marginal values.}

\item{v}{A vector of uniform marginal values.}

\item{series}{An index vector that subsets the full marginal values
\code{u, v}}

\item{fam1}{An integer representing the family of the component copula applied
to u. This is passed to \code{\link{cop_static}}}

\item{fam2}{An integer representing the family of the component copula applied
to v. This is passed to \code{\link{cop_static}}}

\item{date_names}{Optional vector of date names}

\item{parallel}{Logical switch to run the breakpoint search in parallel.}

\item{ncores}{Integer value specifying the number of cores to use}

\item{cluster}{If run in parallel, cluster is passed from a call to
\code{\link{BPfit}}}
}
\value{
A list of: \describe{
  \item{index}{Index, with resepect to the \strong{full} marginal series, of
    largest breakpoint test statistic}
  \item{Date}{If \code{date_names} was provided, the date that corresponds
    with \code{index}}
  \item{sqrt(Z)}{The value of the largest breakpoint test statistic}
  \item{p-value}{The approximated p-value of the breakpoint test statistic}
  \item{N-size}{The number of observations in \code{series} where the function
    searched for a breakpoint}
  }
}
\description{
For the sequential breakpoint model, search for a breakpoint in a subset of 
values for two given marginal distributions#'
}

