% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynohelpers.R
\name{BreakPointLL}
\alias{BreakPointLL}
\title{For the sequential breakpoint model, calculate a generalized likelihood ratio
test}
\usage{
BreakPointLL(u, v, t, fam1, fam2)
}
\arguments{
\item{u}{A vector of uniform marginal values.}

\item{v}{A vector of uniform marginal values.}

\item{t}{An integer indicating where to seperate \code{u} and \code{v} into
pre and post regimes}

\item{fam1}{An integer representing the family of the component copula applied
to u. This is passed to \code{\link{cop_static}}}

\item{fam2}{An integer representing the family of the component copula applied
to v. This is passed to \code{\link{cop_static}}}
}
\value{
A list of: \enumerate{
  \item Pre-regime log-likelihood value
  \item Pre-regime copula parameters
  \item Post-regime log-likeihood value
  \item Post-regime copula parameters
  }
}
\description{
For the sequential breakpoint model, calculate a generalized likelihood ratio
test
}

