% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynohelpers.R
\name{BreakPointTestStatistic}
\alias{BreakPointTestStatistic}
\title{For the sequential breakpoint model, calculate a generalized likelihood ratio
test}
\usage{
BreakPointTestStatistic(t, u, v, fam1, fam2, Full)
}
\arguments{
\item{t}{An integer indicating where to seperate \code{u} and \code{v} into
pre and post regimes}

\item{u}{A vector of uniform marginal values.}

\item{v}{A vector of uniform marginal values.}

\item{fam1}{An integer representing the family of the component copula applied
to u. This is passed to \code{\link{cop_static}}}

\item{fam2}{An integer representing the family of the component copula applied
to v. This is passed to \code{\link{cop_static}}}

\item{Full}{A model from \code{\link{cop_static}} providing the full series'
log-likelihood value}

\item{t}{An integer indexing where to seperate \code{u} and \code{v} into
pre and post regimes}
}
\value{
The test statistic for the generalized likelihood ratio test
}
\description{
For the sequential breakpoint model, calculate a generalized likelihood ratio
test
}

