% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynofits.R
\name{coef.markovCopula}
\alias{coef.markovCopula}
\title{Extract the coefficients from the markov switching model.}
\usage{
\method{coef}{markovCopula}(obj)
}
\arguments{
\item{Obj}{An object of class markovCopula}
}
\description{
Extract the coefficients from the markov switching model.
}

