% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynofits.R
\name{coef.seqBreakPoint}
\alias{coef.seqBreakPoint}
\title{Extract the coefficients from the sequential break point model.}
\usage{
\method{coef}{seqBreakPoint}(obj)
}
\arguments{
\item{Obj}{An object of class seqBreakPoint}
}
\description{
Extract the coefficients from the sequential break point model.
}

