% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynofits.R
\name{coef.smoothTransCopula}
\alias{coef.smoothTransCopula}
\title{Extract the coefficients from the smooth transition model.}
\usage{
\method{coef}{smoothTransCopula}(obj)
}
\arguments{
\item{Obj}{An object of class smoothTransCopula}
}
\description{
Extract the coefficients from the smooth transition model.
}

