% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynohelpers.R
\name{cop_cdf}
\alias{cop_cdf}
\title{Calculate distribution of two-component mixture copula}
\usage{
cop_cdf(theta, u, v, fam1, fam2)
}
\arguments{
\item{theta}{A vector of length 3. The first item is the (single) parameter
for the first component copula, the second item is the (single) parameter
for the second component copula, and the third is the copula weight.}

\item{u}{A vector of uniform marginal values.}

\item{v}{A vector of uniform marginal values.}

\item{fam1}{An integer representing the family of the component copula applied
to u. See Details.}

\item{fam2}{An integer representing the family of the component copula applied
to v. See Details.}
}
\value{
A numeric vector of the same length as \code{u} and \code{v} whose
   values are the distribution (cdf) of the given mixture copula
}
\description{
Calculate distribution of two-component mixture copula
}

