% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynohelpers.R
\name{cop_hessian}
\alias{cop_hessian}
\title{For the sequential breakpoint model, calculate copula hessians for var-cov
estimation}
\usage{
cop_hessian(u, v, theta, f1, f2)
}
\arguments{
\item{u}{A vector of uniform marginal values.}

\item{v}{A vector of uniform marginal values.}

\item{theta}{A vector of copula parameters}

\item{f1}{An integer representing the family of the component copula applied
to u. This is passed to \code{\link{cop_static}}}

\item{f2}{An integer representing the family of the component copula applied
to v. This is passed to \code{\link{cop_static}}}
}
\value{
An nr x nr matrix of gradient contributions
}
\description{
For the sequential breakpoint model, calculate copula hessians for var-cov
estimation
}

