% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynohelpers.R
\name{cop_llh}
\alias{cop_llh}
\title{Convienence function that returns the sum of the negative log-likelihood of
a mixture copula}
\usage{
cop_llh(theta, u, v, fam1, fam2)
}
\arguments{
\item{theta}{A vector of length 3. The first item is the (single) parameter
for the first component copula, the second item is the (single) parameter
for the second component copula, and the third is the copula weight.}

\item{u}{A vector of uniform marginal values.}

\item{v}{A vector of uniform marginal values.}

\item{fam1}{An integer representing the family of the component copula applied
to u. See Details.}

\item{fam2}{An integer representing the family of the component copula applied
to v. See Details.}
}
\value{
The negative log-likelhood value
}
\description{
Convienence function that returns the sum of the negative log-likelihood of
a mixture copula
}

