% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynohelpers.R
\name{cop_static}
\alias{cop_static}
\title{Estimates a static copula}
\usage{
cop_static(u, v, fam1, fam2 = NULL)
}
\arguments{
\item{u}{A vector of uniform marginal values.}

\item{v}{A vector of uniform marginal values.}

\item{fam1}{An integer representing the family of the component copula applied
to u. See Details.}

\item{fam2}{An integer representing the family of the component copula applied
to v. See Details.}
}
\value{
A list of: \enumerate{
  \item the log-likelihood value
  \item the estimated copula parameters
  }
}
\description{
Estimates a static copula
}

