% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynohelpers.R
\name{copula_optim}
\alias{copula_optim}
\title{For the markov switching model, given the parameters that define the
markov transition matrix, complete the maximization step of the E-M algorithm
by optimizing the copula parameters}
\usage{
copula_optim(parm, x, y, family)
}
\arguments{
\item{parm}{A vector of copula parameters}

\item{x}{A vector of uniform marginal values.}

\item{y}{A vector of uniform marginal values.}

\item{family}{Numeric vector of length one or two indicating the copula type
and family.}
}
\value{
A list of: \describe{
  \item{par}{Estimated copula parameters}
  \item{solver}{Output from the call to \code{\link[stats]{optim}}}
  \item{EMstep}{Good to know for further development}
  }
}
\description{
For the markov switching model, given the parameters that define the
markov transition matrix, complete the maximization step of the E-M algorithm
by optimizing the copula parameters
}

