% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynohelpers.R
\name{dependence_measures}
\alias{dependence_measures}
\title{Calculates Kendall's tau, upper tail dependence, and lower tail dependence
for a given copula.}
\usage{
dependence_measures(pars, fam1, fam2 = NULL)
}
\arguments{
\item{pars}{A vector of relevant parameter values}

\item{fam1}{An integer representing the family of the first component copula.}

\item{fam2}{An integer representing the family of the second component copula.
Defaults to \code{NULL}}
}
\value{
A list of: \enumerate{
  \item upper and lower tail-dependence values
  \item Kendall's tau
  }
}
\description{
Calculates Kendall's tau, upper tail dependence, and lower tail dependence
for a given copula.
}
\details{
Providing only \code{fam1} indicates that the user wants the
  dependence values for a single coplula. Providing both \code{fam1} and
  \code{fam2} will return the dependence measures for a mixture copula.
}

