% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynohelpers.R
\name{gaussian_pdf}
\alias{gaussian_pdf}
\title{Vectorized version of the Gaussian copula density}
\usage{
gaussian_pdf(u, v, rho)
}
\arguments{
\item{u}{A vector of uniform marginal values.}

\item{v}{A vector of uniform marginal values.}

\item{rho}{Correlation parameter}
}
\value{
A vector the same length as \code{u} of density values
}
\description{
Vectorized version of the Gaussian copula density
}

