% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynohelpers.R
\name{gumbel_pdf}
\alias{gumbel_pdf}
\title{Vectorized version of the Gumbel copula density}
\usage{
gumbel_pdf(u, v, gamma)
}
\arguments{
\item{u}{A vector of uniform marginal values.}

\item{v}{A vector of uniform marginal values.}

\item{gamma}{Gumbel copula parameter}
}
\value{
A vector the same length as \code{u} of density values
}
\description{
Vectorized version of the Gumbel copula density
}

