% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynofits.R
\name{logLik.markovCopula}
\alias{logLik.markovCopula}
\title{Extract the log-likelihoods from the markov switching model.}
\usage{
\method{logLik}{markovCopula}(obj)
}
\arguments{
\item{Obj}{An object of class markovCopula}
}
\description{
Extract the log-likelihoods from the markov switching model.
}

