% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynofits.R
\name{logLik.seqBreakPoint}
\alias{logLik.seqBreakPoint}
\title{Extract the log-likelihoods from the sequential break point model}
\usage{
\method{logLik}{seqBreakPoint}(obj)
}
\arguments{
\item{Obj}{An object of class seqBreakPoint}
}
\description{
Extract the log-likelihoods from the sequential break point model
}

