% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynofits.R
\name{logLik.smoothTransCopula}
\alias{logLik.smoothTransCopula}
\title{Extract the log-likelihoods from the smooth transition model.}
\usage{
\method{logLik}{smoothTransCopula}(obj)
}
\arguments{
\item{Obj}{An object of class smoothTransCopula}
}
\description{
Extract the log-likelihoods from the smooth transition model.
}

