% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynohelpers.R
\name{markov_llh}
\alias{markov_llh}
\title{For the markov switching model, run the algorithm that produces the model's
log-likelihood.}
\usage{
markov_llh(P, FF, F_0)
}
\arguments{
\item{P}{A markov transition matrix for nr regimes}

\item{FF}{An nr x n matrix where each row represents the densities for each
of the state dependent regimes}

\item{F_0}{An nr x 1 regime of initial regime probabilities}
}
\value{
A list of: \describe{
  \item{logL}{the log-likelihood of the model}
  \item{Xi_t_t}{Xi_t_t}
  \item{Xi_t1_t}{Xi_t1_t}
  }
}
\description{
For the markov switching model, run the algorithm that produces the model's
log-likelihood.
}
\references{
Hamilton (1994) pg ___, eq [22.4.5] and [22.4.5]
}

