% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynohelpers.R
\name{markov_optim}
\alias{markov_optim}
\title{For the markov switching model, given the copula parameters, start the
expectation step of the E-M algorithm by optimizing the markov transition
matrix}
\usage{
markov_optim(parm, x, y, family)
}
\arguments{
\item{parm}{A vector of copula parameters.}

\item{x}{A vector of uniform marginal values.}

\item{y}{A vector of uniform marginal values.}

\item{family}{Numeric vector of length one or two indicating the copula type
and family.}
}
\value{
A list of: \describe{
  \item{par}{Estimated transition parameters and initial state}
  \item{solver}{Output from the call to \code{\link[stats]{optim}}}
  \item{EMstep}{Good to know for further development}
  }
}
\description{
For the markov switching model, given the copula parameters, start the
expectation step of the E-M algorithm by optimizing the markov transition
matrix
}

