% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynohelpers.R
\name{ms_boundary}
\alias{ms_boundary}
\title{For the markov switching model, create boundaries with respect to copula
model to pass on to the optimization function}
\usage{
ms_boundary(fam, k)
}
\arguments{
\item{fam}{Numeric vector of length one or two indicating the copula type and
family.}

\item{k}{Number of regimes}
}
\description{
For the markov switching model, create boundaries with respect to copula
model to pass on to the optimization function
}

