% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynohelpers.R
\name{ms_cop_vpdf}
\alias{ms_cop_vpdf}
\title{For the markov switching model, create function factory that returns a
copula density function conditioning on given copula paramaters}
\usage{
ms_cop_vpdf(fam, x, y)
}
\arguments{
\item{fam}{Numeric vector of length one or two indicating the copula type and
family.}

\item{x}{A vector of uniform marginal values.}

\item{y}{A vector of uniform marginal values.}
}
\description{
For the markov switching model, create function factory that returns a
copula density function conditioning on given copula paramaters
}

