% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynohelpers.R
\name{ms_pnames}
\alias{ms_pnames}
\title{For the markov switching model, return copula names}
\usage{
ms_pnames(x)
}
\arguments{
\item{x}{Integer indicating copula family}
}
\value{
A vector of parameter names
}
\description{
For the markov switching model, return copula names
}

