% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynohelpers.R
\name{numerical_Ktau}
\alias{numerical_Ktau}
\title{For a mixture copula, estimate Kendall's tau using numerical integration}
\usage{
numerical_Ktau(theta, fam1, fam2)
}
\arguments{
\item{theta}{A vector of length 3. The first item is the (single) parameter
for the first component copula, the second item is the (single) parameter
for the second component copula, and the third is the copula weight.}

\item{fam1}{An integer representing the family of the first component copula.}

\item{fam1}{An integer representing the family of the second component copula.}
}
\value{
Kendall's tau for the given mixture copula
}
\description{
For a mixture copula, estimate Kendall's tau using numerical integration
}
\references{
Nelson (2006) pg ___, eq ___
}

