% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynofits.R
\name{plot.markovCopula}
\alias{plot.markovCopula}
\title{Plot of the markov switching model.}
\usage{
\method{plot}{markovCopula}(obj)
}
\arguments{
\item{Obj}{An object of class markovCopula}
}
\description{
Plot of the markov switching model.
}

