% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynofits.R
\name{plot.seqBreakPoint}
\alias{plot.seqBreakPoint}
\title{Plot of the sequential break point model.}
\usage{
\method{plot}{seqBreakPoint}(obj)
}
\arguments{
\item{Obj}{An object of class seqBreakPoint}
}
\description{
Plot of the sequential break point model.
}

