% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynofits.R
\name{plot.smoothTransCopula}
\alias{plot.smoothTransCopula}
\title{Plot of the smooth transition model.}
\usage{
\method{plot}{smoothTransCopula}(obj)
}
\arguments{
\item{Obj}{An object of class smoothTransCopula}
}
\description{
Plot of the smooth transition model.
}

