% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynohelpers.R
\name{Sclayton_pdf}
\alias{Sclayton_pdf}
\title{Vectorized version of the Survival Clayton copula density}
\usage{
Sclayton_pdf(u, v, alpha)
}
\arguments{
\item{u}{A vector of uniform marginal values.}

\item{v}{A vector of uniform marginal values.}

\item{alpha}{Clayton copula parameter}
}
\value{
A vector the same length as \code{u} of density values
}
\description{
Vectorized version of the Survival Clayton copula density
}

