% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynohelpers.R
\name{Sfrank_pdf}
\alias{Sfrank_pdf}
\title{Vectorized version of the Survival Frank copula density}
\usage{
Sfrank_pdf(u, v, delta)
}
\arguments{
\item{u}{A vector of uniform marginal values.}

\item{v}{A vector of uniform marginal values.}

\item{delta}{Frank copula parameter}
}
\value{
A vector the same length as \code{u} of density values
}
\description{
Vectorized version of the Survival Frank copula density
}

