% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynohelpers.R
\name{smooth_inf}
\alias{smooth_inf}
\title{For the markov switching model, an implementation of Kim's smooth inference
alogrithm utilizes the entire series of data to obtain improved inference for
what regime the model occupies at each unit of time}
\usage{
smooth_inf(P, Xi_t_t, Xi_t1_t)
}
\arguments{
\item{P}{A markov transition matrix for nr regimes}

\item{Xi_t_t}{An nr x 1 matrix}

\item{Xi_t1_t}{An nr x 1 matrix}
}
\value{
An nr x n matrix whose columns are the units of time for the time
  series and each row is the probability of being in that regime at that time
}
\description{
For the markov switching model, an implementation of Kim's smooth inference
alogrithm utilizes the entire series of data to obtain improved inference for
what regime the model occupies at each unit of time
}
\references{
Hamilton (1994) pg ___ eq [22.4.14]
}

