% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynohelpers.R
\name{smooth_llh}
\alias{smooth_llh}
\title{For the smooth transition model, calculate the negative log-likelihood value.}
\usage{
smooth_llh(x, y, M, family)
}
\arguments{
\item{x}{A vector of uniform marginal values.}

\item{y}{A vector of uniform marginal values.}

\item{M}{An nr x np matrix whose columns contain the (np) parameters for each
of the nr regimes}

\item{family}{Numeric vector of length one or two indicating the copula type
and family.}
}
\value{
The sum of the negalitve log-likelhood
}
\description{
For the smooth transition model, calculate the negative log-likelihood value.
}

