% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynohelpers.R
\name{smooth_parm}
\alias{smooth_parm}
\title{For the smooth transition model, produce the smooth transition values of a
single parameter}
\usage{
smooth_parm(parm, St, g, c, gsd)
}
\arguments{
\item{parm}{A parameter vector containing a value for each regime}

\item{St}{A vector the same length as the time series whose value is defined
as \code{cumsum(rep(1, N)) \ N}.}

\item{g}{A numeric value controlling the pace of transition between states}

\item{c}{A numeric value in the unit interval locating the inflection point
of the smooth transition}

\item{gsd}{The standard deviation of \code{St}. Combines with the parameter
\code{g} to control the pace of the transition between states}
}
\value{
A vector with the smooth transition values for a copula parameter
}
\description{
For the smooth transition model, produce the smooth transition values of a
single parameter
}

