% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynohelpers.R
\name{st_cop_vpdf}
\alias{st_cop_vpdf}
\title{For the smooth transition model, create function factory that returns a
copula density function conditioning on given copula paramaters}
\usage{
st_cop_vpdf(family, x, y)
}
\arguments{
\item{family}{Numeric vector of length one or two indicating the copula type
and family.}

\item{x}{A vector of uniform marginal values.}

\item{y}{A vector of uniform marginal values.}
}
\value{
A density function for a copula for arbitrary parameter(s)
}
\description{
For the smooth transition model, create function factory that returns a
copula density function conditioning on given copula paramaters
}

