% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynohelpers.R
\name{starting_guess}
\alias{starting_guess}
\title{Provide semi-smart starting values for dynamic copulas.}
\usage{
starting_guess(u, v, fam, k)
}
\arguments{
\item{u}{A vector of uniform marginal values.}

\item{v}{A vector of uniform marginal values.}

\item{fam}{A vector of length one or two indicating which copula family to
find initial values. A vector of length two indicates a mixture copula}

\item{k}{Integer indicating how many segments to disjoint sets to break (u, v)
into.}
}
\value{
A list of length \code{k} with estimated copula parameters
}
\description{
Provide semi-smart starting values for dynamic copulas.
}
\details{
This function breaks the given fixed marginals into \code{k} disjoint
  but adjacent sets of equal size. It then returns the estimated parameters of
  a static copula governed by copula family \code{fam} for each of the disjoint
  sets.
}

