% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynofits.R
\name{summary.markovCopula}
\alias{summary.markovCopula}
\title{Summary of the markov switching model.}
\usage{
\method{summary}{markovCopula}(obj)
}
\arguments{
\item{Obj}{An object of class markovCopula}
}
\description{
Summary of the markov switching model.
}

