% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynofits.R
\name{summary.seqBreakPoint}
\alias{summary.seqBreakPoint}
\title{Summary of the sequential break point model.}
\usage{
\method{summary}{seqBreakPoint}(obj)
}
\arguments{
\item{Obj}{An object of class seqBreakPoint}
}
\description{
Summary of the sequential break point model.
}

