% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynofits.R
\name{summary.smoothTransCopula}
\alias{summary.smoothTransCopula}
\title{Summary of the smooth transition model.}
\usage{
\method{summary}{smoothTransCopula}(obj)
}
\arguments{
\item{Obj}{An object of class smoothTransCopula}
}
\description{
Summary of the smooth transition model.
}

