% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynohelpers.R
\name{tCop_pdf}
\alias{tCop_pdf}
\title{Vectorized version of the Student-t copula density}
\usage{
tCop_pdf(u, v, rho, nu)
}
\arguments{
\item{u}{A vector of uniform marginal values.}

\item{v}{A vector of uniform marginal values.}

\item{rho}{Correlation parameter}

\item{nu}{Degrees of Freedom parameter}
}
\value{
A vector the same length as \code{u} of density values
}
\description{
Vectorized version of the Student-t copula density
}

