% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynohelpers.R
\name{transform_parm}
\alias{transform_parm}
\title{For the smooth transition model, produce the smooth transition values of a
single parameter}
\usage{
transform_parm(parm, family, k, St, gsd)
}
\arguments{
\item{parm}{A vector containing all the parameters of the smooth transition
model}

\item{family}{Numeric vector of length one or two indicating the copula type
and family.}

\item{k}{An integer indicating the number of regimes}

\item{St}{A vector the same length as the time series whose value is defined
as \code{cumsum(rep(1, N)) \ N}.}

\item{gsd}{The standard deviation of \code{St}. Combines with the parameter
\code{g} to control the pace of the transition between states}
}
\value{
An n x np matrix whose columns are the smooth transition values of
  the copula model. For a mixture copula, the (single) weight parameter is
  included as an attribute.
}
\description{
For the smooth transition model, produce the smooth transition values of a
single parameter
}

