% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynofits.R
\name{MSfit}
\alias{MSfit}
\title{Markov-switching copula model.}
\usage{
MSfit(x, y, family = list(1, 1), initValues, tol = 1e-05)
}
\arguments{
\item{x}{A numeric vector of uniform marginal values.}

\item{y}{A numeric vector of uniform marginal values.}

\item{family}{A list of integers specifying the family of the copula to use
in each regime.}

\item{initValues}{Sometimes optional numeric vector of starting values. See
Details.}

\item{tol}{A numeric value specifying the convergence tolerance of the model.
Specifically, the model reaches convergene if the difference in likelihood
from successive models falls below \code{tol}. Defaults to \code{1e-5}}
}
\value{
\code{MSfit} returns an S3 object of \code{\link[base]{class}}
  \code{markovCopula}.
  
  The summary, plot, coef, and logLik functions will, repectively, print a 
  summarization of the output, a plot of dependence measures, extract model
  parameters, and extract the log-likelihood values.
  
  An object of class \code{markovCopula} has the following components: 
  
  \tabular{ll}{
    \code{log.likelihood} \tab log-likelihood value for the regime \cr
    \code{pars} \tab a vector of coefficients for the copula \cr
    \code{N} \tab the length of the time-series \cr
    \code{solver} \tab the final output from \code{\link[stats]{optim}} \cr
    \code{regime.inference} \tab the model's condition density, conditional
      probability, conditional forecasts, and the smoothed probabilities \cr
    \code{copula} \tab details of the estimated copulas in each regime \cr
    \code{transition} \tab the transition matrix and initial regime vector \cr
    \code{nregimes} \tab the number of regimes in the model \cr
  }
}
\description{
\code{MSfit} estimates a markov switching copula on a bivariate time series
  of uniform marginal distributions.
}
\details{
For \code{initValues}, if the same copula family is used for each regime, no initial values
  need to be supplied. If the user wants different copula families estimated
  in different regimes, initValues need to be supplied. For a model with K
  regimes, the order of the values should be provided as follows:
  \enumerate{
    \item Copula parameters in the order they appear in \code{family}
    \item K * (K - 1) transition variables: \eqn{p_{1,1},...,p_{1,k-1},p_{2,1},...,p_{2,k-1},...,p_{k,k-1}}
    \item K - 1 initial state parameters: \eqn{p_{0,1},...,p_{0,k-1} }
  }
}

